Split tests (VmTest and CollectorTest)
